feat: watsonx-ai modules
