+++ pom.xml
@@ -2,17 +2,21 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.4.5</version>
-        <relativePath /> <!-- lookup parent from repository -->
-    </parent>
+
+    <groupId>io.github.springaicommunity</groupId>
     <artifactId>spring-ai-watsonx-ai</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-    <name>spring-ai-watsonx-ai</name>
+    <version>1.1.0-SNAPSHOT</version>
+    <packaging>pom</packaging>
+
+    <modules>
+        <module>spring-ai-autoconfigure-model-watsonx-ai</module>
+        <module>spring-ai-starter-model-watsonx-ai</module>
+        <module>watsonx-ai-core</module>
+    </modules>
+
+    <name>Spring AI watsonx.ai</name>
     <description>spring-ai-watsonx-ai</description>
-    <url/>
+    <url>https://github.com/spring-ai-community/spring-ai-watsonx-ai</url>
 
     <licenses>
         <license>
@@ -46,10 +50,19 @@
         <spotless.version>2.43.0</spotless.version>
         <google-java-format.version>1.17.0</google-java-format.version>
         <spring-ai.version>1.0.0</spring-ai.version>
+        <spring-boot.version>3.5.3</spring-boot.version>
     </properties>
 
     <dependencyManagement>
         <dependencies>
+            <dependency>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-dependencies</artifactId>
+                <version>${spring-boot.version}</version>
+                <type>pom</type>
+                <scope>import</scope>
+            </dependency>
+
             <dependency>
                 <groupId>org.springframework.ai</groupId>
                 <artifactId>spring-ai-bom</artifactId>
@@ -57,41 +70,28 @@
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
+
+            <dependency>
+                <groupId>io.github.springaicommunity</groupId>
+                <artifactId>watsonx-ai-core</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>io.github.springaicommunity</groupId>
+                <artifactId>spring-ai-autoconfigure-model-watsonx-ai</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>io.github.springaicommunity</groupId>
+                <artifactId>spring-ai-starter-model-watsonx-ai</artifactId>
+                <version>${project.version}</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
 
-    <dependencies>
-        <dependency>
-            <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-client-chat</artifactId>
-            <version>${spring-ai.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-retry</artifactId>
-            <version>${spring-ai.version}</version>
-        </dependency>
-        
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-webflux</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
-        </dependency>
-        
-        <!-- IBM Cloud SDK for IAM authentication -->
-        <dependency>
-            <groupId>com.ibm.cloud</groupId>
-            <artifactId>sdk-core</artifactId>
-            <version>${ibm.sdk.version}</version>
-        </dependency>
-    </dependencies>
-
+   
     <build>
         <plugins>
             <plugin>
