+++ spring-ai-autoconfigure-model-watsonx-ai/src/main/java/io/github/springaicommunity/watsonx/autconfigure/chat/WatsonxAiChatAutoConfiguration.java
@@ -1,0 +1,3 @@
+package io.github.springaicommunity.watsonx.autconfigure.chat;
+
+public class WatsonxAiChatAutoConfiguration {}
