+++ spring-ai-starter-model-watsonx-ai/pom.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+          <groupId>io.github.springaicommunity</groupId>
+          <artifactId>spring-ai-watsonx-ai</artifactId>
+          <version>1.1.0-SNAPSHOT</version>
+          <relativePath>../pom.xml</relativePath>
+    </parent>
+
+    <artifactId>spring-ai-starter-model-watsonx-ai</artifactId>
+    <packaging>jar</packaging>
+    <name>Spring AI watsonx.ai Starter</name>
+    <description>Spring AI watsonx.ai Spring Boot Starter</description>
+
+    <properties>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>io.github.springaicommunity</groupId>
+            <artifactId>spring-ai-autoconfigure-model-watsonx-ai</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>io.github.springaicommunity</groupId>
+            <artifactId>watsonx-ai-core</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.ai</groupId>
+            <artifactId>spring-ai-autoconfigure-model-chat-client</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.ai</groupId>
+            <artifactId>spring-ai-retry</artifactId>
+        </dependency>
+        
+        <dependency>
+            <groupId>org.springframework.ai</groupId>
+            <artifactId>spring-ai-autoconfigure-retry</artifactId>
+        </dependency>
+    </dependencies>
+</project>
