--- src/main/java/org/springframewor/ai/watsonx/SpringAiWatsonxAiApplication.java
@@ -1,12 +1,0 @@
-package org.springframewor.ai.watsonx;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpringAiWatsonxAiApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(SpringAiWatsonxAiApplication.class, args);
-  }
-}
