--- src/main/java/org/springframewor/ai/watsonx/aot/WatsonxAiRuntimeHints.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2025 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframewor.ai.watsonx.aot;
-
-import org.springframework.ai.aot.AiRuntimeHints;
-import org.springframework.aot.hint.MemberCategory;
-import org.springframework.aot.hint.RuntimeHints;
-import org.springframework.aot.hint.RuntimeHintsRegistrar;
-import org.springframework.lang.NonNull;
-import org.springframework.lang.Nullable;
-
-/**
- * {@class WatsonxAiRuntimeHints} is responsible for registering runtime hints for watsonx API
- * classes.
- *
- * <p>TODO: add more explanation
- *
- * @author Tristan Mahinay
- * @since 1.1.0-SNAPSHOT
- */
-public class WatsonxAiRuntimeHints implements RuntimeHintsRegistrar {
-
-  @Override
-  public void registerHints(@NonNull RuntimeHints hints, @Nullable ClassLoader classLoader) {
-    var memberCategories = MemberCategory.values();
-
-    for (var typedReference :
-        AiRuntimeHints.findJsonAnnotatedClassesInPackage("org.springframework.ai.openai")) {
-      hints.reflection().registerType(typedReference, memberCategories);
-    }
-  }
-}
