--- src/main/java/org/springframewor/ai/watsonx/extraction/WatsonxAiTextExtractionRequest.java
@@ -1,3 +1,0 @@
-package org.springframewor.ai.watsonx.extraction;
-
-public class WatsonxAiTextExtractionRequest {}
