--- src/main/java/org/springframewor/ai/watsonx/extraction/WatsonxAiTextExtractionResponse.java
@@ -1,3 +1,0 @@
-package org.springframewor.ai.watsonx.extraction;
-
-public class WatsonxAiTextExtractionResponse {}
