+++ watsonx-ai-core/src/main/java/io/github/springaicommunity/watsonx/SpringAiWatsonxAiApplication.java
@@ -1,0 +1,12 @@
+package io.github.springaicommunity.watsonx;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringAiWatsonxAiApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(SpringAiWatsonxAiApplication.class, args);
+  }
+}
