+++ watsonx-ai-core/src/main/java/io/github/springaicommunity/watsonx/chat/WatsonxAiChatModel.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2025 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.github.springaicommunity.watsonx.chat;
+
+import org.springframework.ai.chat.model.ChatModel;
+import org.springframework.ai.chat.model.ChatResponse;
+import org.springframework.ai.chat.prompt.ChatOptions;
+import org.springframework.ai.chat.prompt.Prompt;
+import reactor.core.publisher.Flux;
+
+/**
+ * {@class ChatModel} and {@class org.springframework.ai.chat.model.StreamingChatModel}
+ * implementation that provides access to watsonx supported language models.
+ *
+ * <p>TODO: add more explanation
+ *
+ * @author Tristan Mahinay
+ * @since 1.1.0-SNAPSHOT
+ */
+public class WatsonxAiChatModel implements ChatModel {
+
+  @Override
+  public ChatResponse call(Prompt prompt) {
+
+    // Implement this for watsonx support.
+    return null;
+  }
+
+  @Override
+  public Flux<ChatResponse> stream(Prompt prompt) {
+
+    // Implement this for watsonx support.
+    return null;
+  }
+
+  @Override
+  public ChatOptions getDefaultOptions() {
+    return ChatOptions.builder().build();
+  }
+}
