+++ watsonx-ai-core/src/main/java/io/github/springaicommunity/watsonx/extraction/WatsonxAiTextExtractionResponse.java
@@ -1,0 +1,3 @@
+package io.github.springaicommunity.watsonx.extraction;
+
+public class WatsonxAiTextExtractionResponse {}
