+++ watsonx-ai-core/src/main/resources/application.properties
