+++ watsonx-ai-core/src/test/java/org/springframewor/ai/watsonx/SpringAiWatsonxAiApplicationTests.java
